align package names in test
