+++ core/amber/src/test/scala/edu/uci/ics/texera/workflow/operators/difference/DifferenceOpExecSpec.scala
@@ -1,26 +1,19 @@
-package edu.uci.ics.texera.unittest.workflow.operators.intersect
+package edu.uci.ics.texera.workflow.operators.difference
 
 import edu.uci.ics.amber.engine.common.InputExhausted
 import edu.uci.ics.amber.engine.common.virtualidentity.{LayerIdentity, LinkIdentity}
-import edu.uci.ics.texera.workflow.common.tuple.schema.{Attribute, AttributeType, Schema}
 import edu.uci.ics.texera.workflow.common.tuple.Tuple
-import edu.uci.ics.texera.workflow.operators.intersect.IntersectOpExec
-import org.scalatest.flatspec.AnyFlatSpec
+import edu.uci.ics.texera.workflow.common.tuple.schema.{Attribute, AttributeType}
 import org.scalatest.BeforeAndAfter
+import org.scalatest.flatspec.AnyFlatSpec
 
 import scala.util.Random
-class IntersectOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
-  var opExec: IntersectOpExec = _
-  var counter: Int = 0
 
-  val tupleSchema: Schema = Schema
-    .newBuilder()
-    .add(new Attribute("field1", AttributeType.STRING))
-    .add(new Attribute("field2", AttributeType.INTEGER))
-    .add(
-      new Attribute("field3", AttributeType.BOOLEAN)
-    )
-    .build()
+class DifferenceOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
+  var linkID1: LinkIdentity = _
+  var linkID2: LinkIdentity = _
+  var opExec: DifferenceOpExec = _
+  var counter: Int = 0
 
   def layerID(): LayerIdentity = {
     counter += 1
@@ -32,7 +25,7 @@ class IntersectOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
   def tuple(): Tuple = {
     counter += 1
     Tuple
-      .newBuilder(tupleSchema)
+      .newBuilder()
       .add(new Attribute("field1", AttributeType.STRING), "hello")
       .add(new Attribute("field2", AttributeType.INTEGER), counter)
       .add(
@@ -43,7 +36,9 @@ class IntersectOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
   }
 
   before {
-    opExec = new IntersectOpExec()
+    linkID1 = linkID()
+    linkID2 = linkID()
+    opExec = new DifferenceOpExec(linkID2)
   }
 
   it should "open" in {
@@ -53,8 +48,6 @@ class IntersectOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
   }
 
   it should "work with basic two input streams with no duplicates" in {
-    val linkID1 = linkID()
-    val linkID2 = linkID()
     opExec.open()
     counter = 0
     val commonTuples = (1 to 10).map(_ => tuple()).toList
@@ -69,33 +62,10 @@ class IntersectOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
     })
 
     val outputTuples: Set[Tuple] = opExec.processTexeraTuple(Right(InputExhausted()), linkID2).toSet
-    assert(outputTuples.equals(commonTuples.slice(5, 8).toSet))
+    assert(
+      outputTuples.equals(commonTuples.slice(0, 5).toSet)
+    )
 
-    opExec.close()
-  }
-
-  it should "work with two random input upstreams" in {
-    val links = (0 to 1).map(_ => linkID()).toList
-    opExec.open()
-    counter = 0
-    val commonTuples = (1 to 10).map(_ => tuple()).toList
-
-    (1 to 10).map(_ => {
-
-      opExec.processTexeraTuple(Left(tuple()), links(Random.nextInt(links.size)))
-      opExec.processTexeraTuple(
-        Left(commonTuples(Random.nextInt(commonTuples.size))),
-        links(Random.nextInt(links.size))
-      )
-    })
-
-    assert(opExec.processTexeraTuple(Right(InputExhausted()), links.head).isEmpty)
-
-    val outputTuples: Set[Tuple] =
-      opExec.processTexeraTuple(Right(InputExhausted()), links(1)).toSet
-    assert(outputTuples.size <= 10)
-    assert(outputTuples.subsetOf(commonTuples.toSet))
-    outputTuples.foreach(tuple => assert(tuple.getField[Int]("field2") <= 10))
     opExec.close()
   }
 
@@ -119,53 +89,47 @@ class IntersectOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
   }
 
   it should "work with one empty input upstream after a data stream" in {
-    val linkID1 = linkID()
-    val linkID2 = linkID()
     opExec.open()
     counter = 0
     val commonTuples = (1 to 10).map(_ => tuple()).toList
 
-    (1 to 100).map(_ => {
-      opExec.processTexeraTuple(Left(tuple()), linkID1)
-      opExec.processTexeraTuple(Left(commonTuples(Random.nextInt(commonTuples.size))), linkID1)
+    (0 to 9).map(i => {
+      opExec.processTexeraTuple(Left(commonTuples(i)), linkID1)
     })
     assert(opExec.processTexeraTuple(Right(InputExhausted()), linkID1).isEmpty)
 
-    assert(opExec.processTexeraTuple(Right(InputExhausted()), linkID2).isEmpty)
+    val outputTuples: Set[Tuple] = opExec.processTexeraTuple(Right(InputExhausted()), linkID2).toSet
+    assert(outputTuples.equals(commonTuples.toSet))
     opExec.close()
   }
 
   it should "work with one empty input upstream after a data stream - other order" in {
-    val linkID1 = linkID()
-    val linkID2 = linkID()
     opExec.open()
     counter = 0
     val commonTuples = (1 to 10).map(_ => tuple()).toList
 
-    (1 to 100).map(_ => {
-      opExec.processTexeraTuple(Left(tuple()), linkID1)
-      opExec.processTexeraTuple(Left(commonTuples(Random.nextInt(commonTuples.size))), linkID1)
+    (0 to 9).map(i => {
+      opExec.processTexeraTuple(Left(commonTuples(i)), linkID2)
     })
     assert(opExec.processTexeraTuple(Right(InputExhausted()), linkID2).isEmpty)
 
-    assert(opExec.processTexeraTuple(Right(InputExhausted()), linkID1).isEmpty)
+    val outputTuples: Set[Tuple] = opExec.processTexeraTuple(Right(InputExhausted()), linkID1).toSet
+    assert(outputTuples.isEmpty)
     opExec.close()
   }
 
   it should "work with one empty input upstream before a data stream" in {
-    val linkID1 = linkID()
-    val linkID2 = linkID()
     opExec.open()
     counter = 0
     val commonTuples = (1 to 10).map(_ => tuple()).toList
 
     assert(opExec.processTexeraTuple(Right(InputExhausted()), linkID2).isEmpty)
-    (1 to 100).map(_ => {
-      opExec.processTexeraTuple(Left(tuple()), linkID1)
-      opExec.processTexeraTuple(Left(commonTuples(Random.nextInt(commonTuples.size))), linkID1)
+    (0 to 9).map(i => {
+      opExec.processTexeraTuple(Left(commonTuples(i)), linkID1)
     })
-    assert(opExec.processTexeraTuple(Right(InputExhausted()), linkID1).isEmpty)
 
+    val outputTuples: Set[Tuple] = opExec.processTexeraTuple(Right(InputExhausted()), linkID1).toSet
+    assert(outputTuples.equals(commonTuples.toSet))
     opExec.close()
   }
 
@@ -176,18 +140,16 @@ class IntersectOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
     counter = 0
     val commonTuples = (1 to 10).map(_ => tuple()).toList
 
-    (1 to 100).map(_ => {
-      opExec.processTexeraTuple(Left(tuple()), linkID1)
-      opExec.processTexeraTuple(Left(commonTuples(Random.nextInt(commonTuples.size))), linkID1)
+    (0 to 5).map(i => {
+      opExec.processTexeraTuple(Left(commonTuples(i)), linkID1)
     })
     assert(opExec.processTexeraTuple(Right(InputExhausted()), linkID2).isEmpty)
-
-    (1 to 100).map(_ => {
-      opExec.processTexeraTuple(Left(tuple()), linkID1)
-      opExec.processTexeraTuple(Left(commonTuples(Random.nextInt(commonTuples.size))), linkID1)
+    (6 to 9).map(i => {
+      opExec.processTexeraTuple(Left(commonTuples(i)), linkID1)
     })
-    assert(opExec.processTexeraTuple(Right(InputExhausted()), linkID1).isEmpty)
 
+    val outputTuples: Set[Tuple] = opExec.processTexeraTuple(Right(InputExhausted()), linkID1).toSet
+    assert(outputTuples.equals(commonTuples.toSet))
     opExec.close()
   }
