+++ core/amber/src/test/scala/edu/uci/ics/texera/workflow/operators/hashjoin/HashJoinOpExecSpec.scala
@@ -1,14 +1,9 @@
-package edu.uci.ics.texera.unittest.workflow.operators.hashjoin
+package edu.uci.ics.texera.workflow.operators.hashjoin
 
 import edu.uci.ics.amber.engine.common.InputExhausted
 import edu.uci.ics.amber.engine.common.virtualidentity.{LayerIdentity, LinkIdentity}
 import edu.uci.ics.texera.workflow.common.tuple.Tuple
-import edu.uci.ics.texera.workflow.common.tuple.schema.{
-  Attribute,
-  AttributeType,
-  OperatorSchemaInfo,
-  Schema
-}
+import edu.uci.ics.texera.workflow.common.tuple.schema.{Attribute, AttributeType, OperatorSchemaInfo, Schema}
 import edu.uci.ics.texera.workflow.operators.hashJoin.{HashJoinOpDesc, HashJoinOpExec}
 import org.scalatest.BeforeAndAfter
 import org.scalatest.flatspec.AnyFlatSpec
