+++ core/amber/src/test/scala/edu/uci/ics/texera/workflow/operators/intersect/IntersectOpExecSpec.scala
@@ -1,17 +1,25 @@
-package edu.uci.ics.texera.unittest.workflow.operators.symmetricDifference
+package edu.uci.ics.texera.workflow.operators.intersect
 
 import edu.uci.ics.amber.engine.common.InputExhausted
 import edu.uci.ics.amber.engine.common.virtualidentity.{LayerIdentity, LinkIdentity}
-import edu.uci.ics.texera.workflow.common.tuple.schema.{Attribute, AttributeType}
 import edu.uci.ics.texera.workflow.common.tuple.Tuple
-import edu.uci.ics.texera.workflow.operators.symmetricDifference.SymmetricDifferenceOpExec
-import org.scalatest.flatspec.AnyFlatSpec
+import edu.uci.ics.texera.workflow.common.tuple.schema.{Attribute, AttributeType, Schema}
 import org.scalatest.BeforeAndAfter
+import org.scalatest.flatspec.AnyFlatSpec
 
 import scala.util.Random
-class SymmetricDifferenceOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
-  var opExec: SymmetricDifferenceOpExec = _
+class IntersectOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
+  var opExec: IntersectOpExec = _
   var counter: Int = 0
+
+  val tupleSchema: Schema = Schema
+    .newBuilder()
+    .add(new Attribute("field1", AttributeType.STRING))
+    .add(new Attribute("field2", AttributeType.INTEGER))
+    .add(
+      new Attribute("field3", AttributeType.BOOLEAN)
+    )
+    .build()
 
   def layerID(): LayerIdentity = {
     counter += 1
@@ -23,7 +31,7 @@ class SymmetricDifferenceOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
   def tuple(): Tuple = {
     counter += 1
     Tuple
-      .newBuilder()
+      .newBuilder(tupleSchema)
       .add(new Attribute("field1", AttributeType.STRING), "hello")
       .add(new Attribute("field2", AttributeType.INTEGER), counter)
       .add(
@@ -34,7 +42,7 @@ class SymmetricDifferenceOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
   }
 
   before {
-    opExec = new SymmetricDifferenceOpExec()
+    opExec = new IntersectOpExec()
   }
 
   it should "open" in {
@@ -60,10 +68,33 @@ class SymmetricDifferenceOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
     })
 
     val outputTuples: Set[Tuple] = opExec.processTexeraTuple(Right(InputExhausted()), linkID2).toSet
-    assert(
-      outputTuples.equals(commonTuples.slice(0, 5).toSet.union(commonTuples.slice(8, 10).toSet))
-    )
+    assert(outputTuples.equals(commonTuples.slice(5, 8).toSet))
 
+    opExec.close()
+  }
+
+  it should "work with two random input upstreams" in {
+    val links = (0 to 1).map(_ => linkID()).toList
+    opExec.open()
+    counter = 0
+    val commonTuples = (1 to 10).map(_ => tuple()).toList
+
+    (1 to 10).map(_ => {
+
+      opExec.processTexeraTuple(Left(tuple()), links(Random.nextInt(links.size)))
+      opExec.processTexeraTuple(
+        Left(commonTuples(Random.nextInt(commonTuples.size))),
+        links(Random.nextInt(links.size))
+      )
+    })
+
+    assert(opExec.processTexeraTuple(Right(InputExhausted()), links.head).isEmpty)
+
+    val outputTuples: Set[Tuple] =
+      opExec.processTexeraTuple(Right(InputExhausted()), links(1)).toSet
+    assert(outputTuples.size <= 10)
+    assert(outputTuples.subsetOf(commonTuples.toSet))
+    outputTuples.foreach(tuple => assert(tuple.getField[Int]("field2") <= 10))
     opExec.close()
   }
 
@@ -93,13 +124,13 @@ class SymmetricDifferenceOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
     counter = 0
     val commonTuples = (1 to 10).map(_ => tuple()).toList
 
-    (0 to 9).map(i => {
-      opExec.processTexeraTuple(Left(commonTuples(i)), linkID1)
+    (1 to 100).map(_ => {
+      opExec.processTexeraTuple(Left(tuple()), linkID1)
+      opExec.processTexeraTuple(Left(commonTuples(Random.nextInt(commonTuples.size))), linkID1)
     })
     assert(opExec.processTexeraTuple(Right(InputExhausted()), linkID1).isEmpty)
 
-    val outputTuples: Set[Tuple] = opExec.processTexeraTuple(Right(InputExhausted()), linkID2).toSet
-    assert(outputTuples.equals(commonTuples.toSet))
+    assert(opExec.processTexeraTuple(Right(InputExhausted()), linkID2).isEmpty)
     opExec.close()
   }
 
@@ -110,13 +141,13 @@ class SymmetricDifferenceOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
     counter = 0
     val commonTuples = (1 to 10).map(_ => tuple()).toList
 
-    (0 to 9).map(i => {
-      opExec.processTexeraTuple(Left(commonTuples(i)), linkID2)
+    (1 to 100).map(_ => {
+      opExec.processTexeraTuple(Left(tuple()), linkID1)
+      opExec.processTexeraTuple(Left(commonTuples(Random.nextInt(commonTuples.size))), linkID1)
     })
     assert(opExec.processTexeraTuple(Right(InputExhausted()), linkID2).isEmpty)
 
-    val outputTuples: Set[Tuple] = opExec.processTexeraTuple(Right(InputExhausted()), linkID1).toSet
-    assert(outputTuples.equals(commonTuples.toSet))
+    assert(opExec.processTexeraTuple(Right(InputExhausted()), linkID1).isEmpty)
     opExec.close()
   }
 
@@ -128,12 +159,12 @@ class SymmetricDifferenceOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
     val commonTuples = (1 to 10).map(_ => tuple()).toList
 
     assert(opExec.processTexeraTuple(Right(InputExhausted()), linkID2).isEmpty)
-    (0 to 9).map(i => {
-      opExec.processTexeraTuple(Left(commonTuples(i)), linkID1)
+    (1 to 100).map(_ => {
+      opExec.processTexeraTuple(Left(tuple()), linkID1)
+      opExec.processTexeraTuple(Left(commonTuples(Random.nextInt(commonTuples.size))), linkID1)
     })
+    assert(opExec.processTexeraTuple(Right(InputExhausted()), linkID1).isEmpty)
 
-    val outputTuples: Set[Tuple] = opExec.processTexeraTuple(Right(InputExhausted()), linkID1).toSet
-    assert(outputTuples.equals(commonTuples.toSet))
     opExec.close()
   }
 
@@ -144,16 +175,18 @@ class SymmetricDifferenceOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
     counter = 0
     val commonTuples = (1 to 10).map(_ => tuple()).toList
 
-    (0 to 5).map(i => {
-      opExec.processTexeraTuple(Left(commonTuples(i)), linkID1)
+    (1 to 100).map(_ => {
+      opExec.processTexeraTuple(Left(tuple()), linkID1)
+      opExec.processTexeraTuple(Left(commonTuples(Random.nextInt(commonTuples.size))), linkID1)
     })
     assert(opExec.processTexeraTuple(Right(InputExhausted()), linkID2).isEmpty)
-    (6 to 9).map(i => {
-      opExec.processTexeraTuple(Left(commonTuples(i)), linkID1)
-    })
 
-    val outputTuples: Set[Tuple] = opExec.processTexeraTuple(Right(InputExhausted()), linkID1).toSet
-    assert(outputTuples.equals(commonTuples.toSet))
+    (1 to 100).map(_ => {
+      opExec.processTexeraTuple(Left(tuple()), linkID1)
+      opExec.processTexeraTuple(Left(commonTuples(Random.nextInt(commonTuples.size))), linkID1)
+    })
+    assert(opExec.processTexeraTuple(Right(InputExhausted()), linkID1).isEmpty)
+
     opExec.close()
   }
