+++ core/amber/src/test/scala/edu/uci/ics/texera/workflow/operators/projection/ProjectionOpDescSpec.scala
@@ -1,9 +1,8 @@
-package edu.uci.ics.texera.unittest.workflow.operators.projection
+package edu.uci.ics.texera.workflow.operators.projection
 
 import edu.uci.ics.texera.workflow.common.tuple.schema.{Attribute, AttributeType, Schema}
-import edu.uci.ics.texera.workflow.operators.projection.ProjectionOpDesc
-import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.BeforeAndAfter
+import org.scalatest.flatspec.AnyFlatSpec
 
 import scala.collection.convert.ImplicitConversions.`list asScalaBuffer`
 class ProjectionOpDescSpec extends AnyFlatSpec with BeforeAndAfter {
