+++ core/amber/src/test/scala/edu/uci/ics/texera/workflow/operators/symmetricDifference/SymmetricDifferenceOpExecSpec.scala
@@ -1,19 +1,15 @@
-package edu.uci.ics.texera.unittest.workflow.operators.difference
+package edu.uci.ics.texera.workflow.operators.symmetricDifference
 
 import edu.uci.ics.amber.engine.common.InputExhausted
 import edu.uci.ics.amber.engine.common.virtualidentity.{LayerIdentity, LinkIdentity}
-import org.scalatest.BeforeAndAfter
-import org.scalatest.flatspec.AnyFlatSpec
 import edu.uci.ics.texera.workflow.common.tuple.Tuple
 import edu.uci.ics.texera.workflow.common.tuple.schema.{Attribute, AttributeType}
-import edu.uci.ics.texera.workflow.operators.difference.DifferenceOpExec
+import org.scalatest.BeforeAndAfter
+import org.scalatest.flatspec.AnyFlatSpec
 
 import scala.util.Random
-
-class DifferenceOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
-  var linkID1: LinkIdentity = _
-  var linkID2: LinkIdentity = _
-  var opExec: DifferenceOpExec = _
+class SymmetricDifferenceOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
+  var opExec: SymmetricDifferenceOpExec = _
   var counter: Int = 0
 
   def layerID(): LayerIdentity = {
@@ -37,9 +33,7 @@ class DifferenceOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
   }
 
   before {
-    linkID1 = linkID()
-    linkID2 = linkID()
-    opExec = new DifferenceOpExec(linkID2)
+    opExec = new SymmetricDifferenceOpExec()
   }
 
   it should "open" in {
@@ -49,6 +43,8 @@ class DifferenceOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
   }
 
   it should "work with basic two input streams with no duplicates" in {
+    val linkID1 = linkID()
+    val linkID2 = linkID()
     opExec.open()
     counter = 0
     val commonTuples = (1 to 10).map(_ => tuple()).toList
@@ -64,7 +60,7 @@ class DifferenceOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
 
     val outputTuples: Set[Tuple] = opExec.processTexeraTuple(Right(InputExhausted()), linkID2).toSet
     assert(
-      outputTuples.equals(commonTuples.slice(0, 5).toSet)
+      outputTuples.equals(commonTuples.slice(0, 5).toSet.union(commonTuples.slice(8, 10).toSet))
     )
 
     opExec.close()
@@ -90,6 +86,8 @@ class DifferenceOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
   }
 
   it should "work with one empty input upstream after a data stream" in {
+    val linkID1 = linkID()
+    val linkID2 = linkID()
     opExec.open()
     counter = 0
     val commonTuples = (1 to 10).map(_ => tuple()).toList
@@ -105,6 +103,8 @@ class DifferenceOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
   }
 
   it should "work with one empty input upstream after a data stream - other order" in {
+    val linkID1 = linkID()
+    val linkID2 = linkID()
     opExec.open()
     counter = 0
     val commonTuples = (1 to 10).map(_ => tuple()).toList
@@ -115,11 +115,13 @@ class DifferenceOpExecSpec extends AnyFlatSpec with BeforeAndAfter {
     assert(opExec.processTexeraTuple(Right(InputExhausted()), linkID2).isEmpty)
 
     val outputTuples: Set[Tuple] = opExec.processTexeraTuple(Right(InputExhausted()), linkID1).toSet
-    assert(outputTuples.isEmpty)
+    assert(outputTuples.equals(commonTuples.toSet))
     opExec.close()
   }
 
   it should "work with one empty input upstream before a data stream" in {
+    val linkID1 = linkID()
+    val linkID2 = linkID()
     opExec.open()
     counter = 0
     val commonTuples = (1 to 10).map(_ => tuple()).toList
