+++ core/amber/src/test/scala/edu/uci/ics/texera/workflow/operators/visualization/htmlviz/HtmlVizOpExecSpec.scala
@@ -1,13 +1,7 @@
-package edu.uci.ics.texera.unittest.workflow.operators.visualization.htmlviz
+package edu.uci.ics.texera.workflow.operators.visualization.htmlviz
 
 import edu.uci.ics.texera.workflow.common.tuple.Tuple
-import edu.uci.ics.texera.workflow.common.tuple.schema.{
-  Attribute,
-  AttributeType,
-  OperatorSchemaInfo,
-  Schema
-}
-import edu.uci.ics.texera.workflow.operators.visualization.htmlviz.{HtmlVizOpDesc, HtmlVizOpExec}
+import edu.uci.ics.texera.workflow.common.tuple.schema.{Attribute, AttributeType, OperatorSchemaInfo, Schema}
 import org.scalatest.BeforeAndAfter
 import org.scalatest.flatspec.AnyFlatSpec
